This is my very first programming project, written in python to create a casino simulator with simple games like slots and roulette.
This is my very first programming project, written in python.

The goal of this project was to apply what I learned about functions, conditional loops, and random number generators to create a really simple game of chance that I called slots and roulette.

I will try to continue to add to this project by hopefully actually creating a slots and roulette game with their real-life rules as well as real-life interfaces. I'll also try to add more games onto this project as well, as a little passion project in my free time as a first-year biomedical engineering student.